Build scripts
Former-commit-id: a593a80079c152ed5fea70271e9490a90a4f5985

+++ build/Wine/update_mclient.sh
@@ -8,5 +8,10 @@ rm windows.7z
 cp -ru build/exe.win32-3.4/* mclient/
 cp -ru dics locale resources mclient.cfg mclient/
 cd mclient && wine mclient.exe
-cd .. && 7z a windows.7z mclient/ && rm -r build mclient
+read -p "Update the archive?" choice
+if [ "$choice" = "y" ] || [ "$choice" = "Y" ]; then
+	cd .. && 7z a windows.7z mclient/ && rm -r build mclient
+	mv -fv $HOME/tmp/ars/windows.7z $HOME/tmp/ars/windows\ \(OLD\).7z
+	mv -v ./windows.7z $HOME/tmp/ars/windows.7z
+fi
 ./clean_up.sh
